fix: check discord server id

+++ src/main/java/io/github/spigotrce/paradiseclientfabric/chatroom/client/DiscordModel.java
@@ -1,4 +1,4 @@
 package io.github.spigotrce.paradiseclientfabric.chatroom.client;
 
-public record DiscordModel(String token, int serverID, boolean autoVerify) {
+public record DiscordModel(String token, long serverID, boolean autoVerify) {
 }
